Fix: null 체크

+++ src/main/java/org/sparta/springlv3/lecture/LectureService.java
@@ -144,6 +144,10 @@ public class LectureService {
 
         Optional<Lecture> lectureOpt = lectureRepository.findById(id);
 
+        if (lectureOpt.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+
         Lecture lecture = lectureOpt.get();
         lectureRepository.delete(lecture);
